Update UI login test initial/password configs
With https://github.com/konveyor/tackle-ui-tests/pull/1543,
it is possible to provide the default admin user password
that needs to be changed on first login.  While the test
configs embed the default username and password, it is useful
to explicitly define them in the workflow so we know
exactly what is being used.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>